Fix webpack warning
