+++ packages/next/server/on-demand-entry-handler.js
@@ -4,7 +4,8 @@ import { join } from 'path'
 import fs from 'fs'
 import promisify from '../lib/promisify'
 import globModule from 'glob'
-import {normalizePagePath, pageNotFoundError} from 'next-server/dist/server/require'
+import {pageNotFoundError} from 'next-server/dist/server/require'
+import {normalizePagePath} from 'next-server/dist/server/normalize-page-path'
 import {createEntry} from '../build/webpack/utils'
 import { ROUTE_NAME_REGEX, IS_BUNDLED_PAGE_REGEX } from 'next-server/constants'
